+++ nflow-engine/src/main/resources/scripts/db/update-1.2.0-x/mysql.update.ddl.sql
@@ -1,2 +1,3 @@
-
 alter table nflow_workflow_definition add definition_sha1 varchar(40) not null default 'n/a';
+alter table nflow_workflow_action add type enum('stateExecution', 'stateExecutionFailed', 'recovery', 'externalChange') not null default 'stateExecution';
+